feat: add tini init system
Using a shell or the executable directly as PID 1 is highly discouraged because of signal handling etc.

A very lightweight init system for containers is tini. Using it as PID 1 will make sure all the boxes are checked for proper runtime compliance of a container.

Tini itself will call the entrypoint script, which will replace itself with the "bw" webserver. This way tini will take care of properly shutting the processes down when the container is terminated.